First steps towards i18n
The metadata is in a collection (_labs) now.
The English pages are in a subfolder called en and use the central metadata.
The original files remain the same in order not to break old urls.
The navigartion is localized as well.
TODO: Only Berlin is done. The rest needs to follow.